create IO hook, move page-specific components to page folders
