--- src/components/Newsletter/index.js
@@ -1,3 +1,0 @@
-import Newsletter from "./Newsletter";
-
-export default Newsletter;
