+++ src/hooks/useOnIntersection.js
@@ -1,0 +1,24 @@
+import React, { useState, useEffect } from "react";
+
+export default function useOnIntersection(ref) {
+  const [isIntersecting, setIsIntersecting] = useState(false);
+
+  useEffect(() => {
+    const options = { treshold: 1 };
+    const callback = (entries) => {
+      entries.forEach((entry) => {
+        if (entry.isIntersecting) {
+          setIsIntersecting(true);
+        }
+      });
+    };
+
+    const observer = new IntersectionObserver(callback, options);
+    observer.observe(ref.current);
+
+    const currentRef = ref.current;
+    return () => observer.unobserve(currentRef);
+  }, []);
+
+  return isIntersecting;
+}
