+++ src/pages/Checkout/Checkout.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../../components/Header/Header";
 import TextSection from "../../components/TextSection";
-import Form from "../../components/Form/CheckoutForm";
+import Form from "./components/Form/CheckoutForm";
 import cartStyles from "../../components/Cart/cart.module.css";
 import checkoutStyles from "./checkout.module.css";
 import SubmitMsg from "../../components/SubmitMsg/SubmitMsg";
