+++ src/pages/Main/components/Newsletter/Newsletter.js
@@ -1,7 +1,7 @@
 import React from "react";
 import newsletterStyles from "./newsletter.module.css";
-import useFormSubmit from "../../hooks/useFormSubmit";
-import textStyles from "../../components/TextSection/text-section.module.css";
+import useFormSubmit from "../../../../hooks/useFormSubmit";
+import textStyles from "../../../../components/TextSection/text-section.module.css";
 
 export default function Newsletter({ submitted }) {
   const initialValue = { email: "" };
