+++ src/pages/Main/components/Newsletter/index.js
@@ -1,0 +1,3 @@
+import Newsletter from ".Newsletter";
+
+export default Newsletter;
