+++ src/pages/Main/components/Slider/Slider.js
