+++ src/pages/Main/components/Slider/custom-slider.module.css
@@ -15,18 +15,18 @@
 }
 
 .slide1 {
-  background-image: url("../../img/slide1.jpg");
+  background-image: url("../../../../img/slide1.jpg");
 }
 
 .slide2 {
-  background-image: url("../../img/slide2.jpg");
+  background-image: url("../../../../img/slide2.jpg");
 }
 
 .slide3 {
-  background-image: url("../../img/slide3.jpg");
+  background-image: url("../../../../img/slide3.jpg");
 }
 
 .slide4 {
-  background-image: url("../../img/slide4.jpg");
+  background-image: url("../../../../img/slide4.jpg");
   background-position: top;
 }
