+++ src/pages/Main/components/Slider/index.js
