+++ src/pages/Main/components/TwoPhotosSection/index.js
