+++ src/pages/Shop/Shop.js
@@ -1,8 +1,8 @@
 import React from "react";
 import TextSection from "../../components/TextSection/TextSection";
 import Header from "../../components/Header/Header";
-import Products from "../../components/ProductsList/Products";
-import Faq from "../../components/FAQ/Faq";
+import ProductList from "./components/ProductsList/Products";
+import Faq from "./components/FAQ/Faq";
 
 const Shop = () => {
   return (
@@ -16,7 +16,7 @@ const Shop = () => {
           subcaption="Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa."
           links={true}
         ></TextSection>
-        <Products />
+        <ProductList />
         <Faq />
       </main>
     </>
