+++ src/pages/Shop/components/FAQ/faq.module.css
