+++ src/pages/Shop/components/FAQ/index.js
