+++ src/pages/Shop/components/ProductsList/Products.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { useProductList } from "../../firebase/api";
+import { useProductList } from "../../../../firebase/api";
 import productStyles from "../../components/ProductsList/product.module.css";
 
 const Products = () => {
