+++ src/pages/Shop/components/ProductsList/index.js
